new try

+++ hashmap.c
@@ -45,7 +45,7 @@ int is_equal(void* key1, void* key2){
 
 
 void insertMap(HashMap * map, char * key, void * value) {
-  long a = hash(key, map -> capacity);
+  unsigned long long a = hash(key , map -> capacity);
   if (map -> buckets[a] != NULL) {
     do {
       a++;
